Add an unsigned to signed cast to surpress a warning when building

+++ sf-cvs/trunk/audacity-src/src/prefs/QualityPrefs.cpp
@@ -229,7 +229,7 @@ bool QualityPrefs::Apply()
 void QualityPrefs::OnSampleRateChoice(wxCommandEvent& evt)
 {
    int sel = mSampleRates->GetSelection();
-   mOtherSampleRate->Enable(sel == mSampleRates->GetCount() - 1);
+   mOtherSampleRate->Enable(sel == (int)mSampleRates->GetCount() - 1);
 }
